Changes on most classes to work from GUI
The Hero now has methods to correctly equip weapons, to level up and change stats, some unused attributes were removed, also a method to search a node in the Tree was added. In monster a lot of changes were made to ensure correct working in Combat, such as some getters and setters in SpritePath in NPC, and etc. Boss constructor was changed to match the game's changes as well.